misc: flatten uniqueness hash

+++ src/dowse.rs
@@ -47,7 +47,7 @@ where P: AsRef<Path>, I: IntoIterator<Item=P> {
 	// Parse out seed paths.
 	let (mut dirs, files, seen) = {
 		let mut files: Vec<PathBuf> = Vec::with_capacity(2048);
-		let mut seen: AHashSet<u128> = AHashSet::with_capacity_and_hasher(2048, AHASH_STATE);
+		let mut seen: AHashSet<u64> = AHashSet::with_capacity_and_hasher(2048, AHASH_STATE);
 
 		let dirs: Vec<ReadDir> = paths.into_iter()
 			.filter_map(|p| resolve_path(PathBuf::from(p.as_ref()), false))
